requirements.txt
streamlit
requests
matplotlib
pillow
